rename FavoritePage.dart to favorite_page.dart
